adding automous code object. This object will drive the rover to a waypoint and avoid obstacles automatically.

+++ src/mainGUI.h
@@ -29,7 +29,7 @@ class MainGUI : public QMainWindow, private Ui::MainGUI
 private:
 	QTcpServer		m_tcpServer;
 	QTcpSocket		*m_tcpSocket;
-	qint64			m_blockSize;
+	quint16			m_blockSize;
 	QDataStream		inStream;
 	QDataStream		outStream;
